Added security to creation, edition and deletion of products

+++ products/views.py
@@ -2,7 +2,7 @@ from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
 from django.db.models import Q
 from django.db.models.functions import Lower
-
+from django.contrib.auth.decorators import login_required
 from .models import Product, Category, Brand
 from .forms import ProductForm
 
@@ -102,8 +102,13 @@ def all_brands(request):
     return render(request, 'products/all_brands.html', context)
 
 
+@login_required
 def add_product(request):
     """ Add a product to the store """
+     if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you are not an admin. NO CAN DO!')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -123,8 +128,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ Edit a product in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you are not an admin. NO CAN DO!')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -147,8 +157,13 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ Delete a product from the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you are not an admin. NO CAN DO!')
+        return redirect(reverse('home'))
+        
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, f'The {product.name} is deleted!')
